refactor routing structure; update route imports, add module name mapping, and remove obsolete App test
